Turn contourf method into a wrapper for xbout.plot.* methods
To make 2d plots, convert the plotfuncs.contourf function into
plot2d_wrapper which takes an xarray plotting method from xarray.plot as
its second argument and uses that method to make a 2d poloidal plot. A
bit of special handling is needed for the color-bars in contour plots.

Methods contourf and pcolormesh for BoutDataArrayAccessor are one-liners
that call plotfuncs.plot2d_wrapper.