Ensure separatrices, target plotted correctly if Dataset transposed
Create coordinates called 'xdim', 'ydim' and 'zdim' that depend only on
'x', 'y' and 'z' respectively, before applying geometry and re-naming
the dimensions. This means that in plot_separatrices() and
plot_targets() we can find renamed dimensions corresponding to 'x' and
'y' even if the Dataset has been transposed so that we cannot rely on
the order of the dimensions.